changes in front and backend both

+++ server/server.js
@@ -9,6 +9,8 @@ import cors from "cors"
 import pg from "pg"
 dotenv.config()
 const app = express();
+app.use(bodyParser.urlencoded({extended:true}));
+app.use(express.json());
 const port = 3000;
 const saltRounds = 10;
 const db=new pg.Client({
@@ -18,11 +20,22 @@ database:process.env.DB_DATABASE,
 password:process.env.DB_PASSWORD,
 port:5432
 })
+app.use(session({
+  secret:process.env.SESSION_SECRET,
+  resave:false,
+  saveUninitialized:true,
+   cookie: {
+      secure: false, // true only when using HTTPS in production
+      maxAge: 1000 * 60 * 60 * 24 // 1 day
+    }
+}))
+app.use(passport.initialize());
+app.use(passport.session());
+
 db.connect();
 
-app.use(bodyParser.urlencoded({extended:true}));
-app.use(express.json());
-app.use(cors({origin: "*" }))
+
+app.use(cors({origin: "http://localhost:5173" , credentials: true, }))
 app.use(express.static("public"));
 app.post("/signup", async(req, res)=>{
   const email= req.body.username;
@@ -45,25 +58,64 @@ app.post("/signup", async(req, res)=>{
     
   }
 })
-app.post("/login", async(req,res)=>{
-  const email = req.body.username;
-  const password = req.body.password;
+// app.post("/login", async(req,res)=>{
+//   const email = req.body.username;
+//   const password = req.body.password;
+//   })
+app.post("/login", (req, res, next) => {
+  passport.authenticate("local", (err, user) => {
+    if (err) return next(err);
+
+    if (!user) {
+      return res.status(401).json({
+        success: false,
+        message: "Invalid credentials",
+      });
+    }
+
+    req.logIn(user, (err) => {
+      if (err) return next(err);
+       console.log("req.user after login:", req.user);
+      res.json({
+        success: true,
+        user,
+      });
+    });
+  })(req, res, next);
+});
+passport.serializeUser((user , cb)=>{
+  cb(null , user.id)
+});
+passport.deserializeUser(async(id, cb)=>{
   try {
-    const result = await db.query(`SELECT * FROM users WHERE email=$1`,[email])
+    const result = await db.query(`SELECT * FROM users WHERE id=$1`, [id]);
+    cb(null , result.rows[0]);
+  } catch (error) {
+    cb(error)
+  }
+});
+  passport.use(
+    new Strategy(async function verify(username,password,cb) {
+        try {
+    const result = await db.query(`SELECT * FROM users WHERE email=$1`,[username])
     if (result.rows.length>0) {
-       const savedHashedPassword = result.rows[0].password_hash
+      const user = result.rows[0]
+       const savedHashedPassword = user.password_hash
        bcrypt.compare(password , savedHashedPassword,(err , result)=>{
         if (err) {
           console.log("error compairing password" , err)
           console.log("error compairing password!")
+           return cb(err);
         } else {
           if (result) {
-            res.status(200).send("login page is ok!!")
+            // res.status(200).send("login page is ok!!")
             console.log("authorized!")
+             return cb(null , user);
           } else {
-            res.status(401).json({
-    message: "Incorrect password"
-});
+//             res.status(401).json({
+//     message: "Incorrect password"
+// });
+ return cb(null , false);
 console.log("Incorrrect Password!")
 
           }
@@ -72,16 +124,20 @@ console.log("Incorrrect Password!")
        })
   
     } else {
-      res.status(404).json({
-    message: "User not found"
-});
-      console.log("User not found!");
+//       res.status(404).json({
+//     message: "User not found"
+// });
+      // console.log("User not found!");
+       return cb("User not found");
     }
    
   } catch (error) {
-    
+     return cb(error);
   }
-})
+    })
+  )
+
+
 app.get("/data", async(req,res)=>{
     const month = req.query.month;
     const year = req.query.year;
@@ -95,16 +151,18 @@ app.get("/data", async(req,res)=>{
   if (month > 0 && year) {
     expenses = await db.query(
       `SELECT * FROM data
-       WHERE EXTRACT(MONTH FROM expense_date)::int = $1
-       AND EXTRACT(YEAR FROM expense_date)::int = $2`,
-      [month, year]
+       WHERE user_id = $1
+       AND EXTRACT(MONTH FROM expense_date)::int = $2
+       AND EXTRACT(YEAR FROM expense_date)::int = $3`,
+      [req.user.id, month, year]
     );
      total = await db.query(
     `SELECT SUM(amount) AS total_expense
      FROM data
-     WHERE EXTRACT(MONTH FROM expense_date)::int = $1
-       AND EXTRACT(YEAR FROM expense_date)::int = $2`,
-    [month, year]
+     WHERE user_id = $1
+     AND EXTRACT(MONTH FROM expense_date)::int = $2
+       AND EXTRACT(YEAR FROM expense_date)::int = $3`,
+    [req.user.id, month, year]
 );
 
 
@@ -113,43 +171,48 @@ app.get("/data", async(req,res)=>{
   } else if (month > 0) {
     expenses = await db.query(
       `SELECT * FROM data
-       WHERE EXTRACT(MONTH FROM expense_date)::int = $1`,
-      [month]
+      WHERE user_id = $1
+       AND EXTRACT(MONTH FROM expense_date)::int = $2`,
+      [ req.user.id , month]
     );
         total = await db.query(
     `SELECT SUM(amount) AS total_expense
      FROM data
-     WHERE EXTRACT(MONTH FROM expense_date)::int = $1`,
-    [month]
+     WHERE user_id = $1
+     AND EXTRACT(MONTH FROM expense_date)::int = $2`,
+    [req.user.id , month]
 );
 
   
   } else if (month === 0 && year) {
     expenses = await db.query(
       `SELECT * FROM data
-       WHERE EXTRACT(YEAR FROM expense_date)::int = $1`,
-      [year]
+      WHERE user_id = $1
+       AND EXTRACT(YEAR FROM expense_date)::int = $2`,
+      [req.user.id , year]
     );
     total = await db.query(
     `SELECT SUM(amount) AS total_expense
      FROM data
-     WHERE EXTRACT(YEAR FROM expense_date)::int = $2`,
-    [year]
+     WHERE user_id = $1
+     AND EXTRACT(YEAR FROM expense_date)::int = $2`,
+    [req.user.id, year]
 );
   
   } else if (year) {
     expenses = await db.query(
       `SELECT * FROM data
-       WHERE EXTRACT(YEAR FROM expense_date)::int = $1`,
-      [year]
+      WHERE user_id = $1
+       AND EXTRACT(YEAR FROM expense_date)::int = $2`,
+      [req.user.id, year]
     );
 
   
   } else {
-    expenses = await db.query(`SELECT * FROM data`);
+    expenses = await db.query(`SELECT * FROM data WHERE user_id = $1`,[req.user.id]);
         total = await db.query(
     `SELECT SUM(amount) AS total_expense
-     FROM data`);
+     FROM data WHERE user_id = $1`,[req.user.id]);
    
   }
 
@@ -173,10 +236,11 @@ app.get("/data", async(req,res)=>{
 // })
 app.post("/data" , async(req ,res)=>{
     const {title , category , amount , expense_date} = req.body
+    const user = req.user.id
     // console.log(req.body)
     try {
-       const result = await db.query("INSERT INTO data(title , category , amount , expense_date) VALUES($1 , $2 , $3 , $4) RETURNING *",
-            [title , category , amount , expense_date]
+       const result = await db.query("INSERT INTO data(user_id ,title , category , amount , expense_date) VALUES($1 , $2 , $3 , $4 , $5) RETURNING *",
+            [user , title , category , amount , expense_date]
         );
         res.json(result.rows[0]);
     } catch (error) {
@@ -188,8 +252,9 @@ app.put("/data/:id" , async(req,res)=>{
 try {
     const {title , category , amount , expense_date} =req.body
     const id = req.params.id;
-    await db.query("UPDATE data SET title = $1 , category = $2 , amount = $3 , expense_date = $4 WHERE id=$5",
-        [title , category , amount , expense_date , id]
+    const user = req.user.id
+    await db.query("UPDATE data SET title = $1 , category = $2 , amount = $3 , expense_date = $4 WHERE id=$5 AND user_id = $6",
+        [title , category , amount , expense_date , id , user]
        
     );
      res.json({message:"Updated successfully!"})
@@ -200,13 +265,15 @@ try {
 app.delete("/data/:id" , async(req, res)=>{
     try {
          const id = req.params.id
-    await db.query("DELETE FROM data WHERE id=$1",[id]);
+         const user = req.user.id
+    await db.query("DELETE FROM data WHERE id=$1 AND user_id = $2",[id , user]);
     res.json({message:"Deleted successfully!"});
     } catch (error) {
         res.sendStatus(500).json({message:"Error deleting the data!"});
     }
    
 });
+
 app.listen(port ,"0.0.0.0" ,  ()=>{
     console.log(`this app is listeninig on ${port}!`);
 });
